requirements.txt
requests
fabric==1.14.0
boto3==1.7.25
paramiko==2.4.1
graphviz==0.8.3
globus_sdk==1.5.0
docker==3.5.0
apache-libcloud==2.3.0
configparser==3.5.0
flask
Flask-API
pycrypto